Compila (putting everithing together)

+++ inc/intarr_utils.h
@@ -6,7 +6,7 @@
 /*   By: glajara- <glajara-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 10:47:42 by glajara-          #+#    #+#             */
-/*   Updated: 2023/10/19 11:39:33 by glajara-         ###   ########.fr       */
+/*   Updated: 2023/10/25 16:27:35 by glajara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdlib.h>
 # include "basic_utils.h"
 
-size_t	intarr_get_size(int *arr);
+int		intarr_get_size(int *arr);
 int		*intarr_dup(int *arr);
 int		*intarr_add(int *arr, int n);
