+++ src/strarr/strarr_add.c
@@ -6,7 +6,7 @@
 /*   By: glajara- <glajara-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 11:42:30 by glajara-          #+#    #+#             */
-/*   Updated: 2023/10/19 10:27:35 by glajara-         ###   ########.fr       */
+/*   Updated: 2023/10/25 16:23:53 by glajara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 char	**strarr2_add(char **arr, char *str)
 {
 	char	**new_arr;
-	size_t	orig_arr_size;
+	int		orig_arr_size;
 	int		i;
 
 	orig_arr_size = strarr2_get_size(arr);
@@ -39,7 +39,7 @@ char	**strarr2_add(char **arr, char *str)
 char	***strarr3_add(char ***arr, char **subarr)
 {
 	char	***new_arr;
-	size_t	orig_arr_size;
+	int		orig_arr_size;
 	int		i;
 
 	orig_arr_size = strarr3_get_size(arr);
